Fixed gemv

+++ test/gtest/sp/gemv.cpp
@@ -69,7 +69,7 @@ TEST(SparseMatrix, ZeroVector) {
   std::vector<std::pair<std::pair<I, I>, T>> base;
   for (int i = 0; i < m; i++) {
     for (int j = 0; j < k; j++) {
-      base.push_back({{i, j}, i + j});
+      base.push_back({{i, j}, static_cast<float>(i + j)});
     }
   }
